+++ interface/RecoObjects/PseudoParticle.h
@@ -17,32 +17,32 @@ typedef std::vector<PseudoParticle> PseudoParticleCollection;
 
 class PseudoParticle {
 protected:
-    float Phi, Eta, Pt, Theta, D0, Charge;
+    double Phi, Eta, Pt, Theta, D0, Charge;
 public:
     PseudoParticle();
-    PseudoParticle(float phi, float eta, float pt, float theta);
+    PseudoParticle(double phi, double eta, double pt, double theta);
     virtual ~PseudoParticle();
-    float pt() const;
-    float eta() const;
-    float phi() const;
-    float theta() const;
-    float charge() const;
-    float d0() const;
+    double pt() const;
+    double eta() const;
+    double phi() const;
+    double theta() const;
+    double charge() const;
+    double d0() const;
 
-    void setPt(float pt);
-    void setEta(float eta);
-    void setPhi(float phi);
-    void setTheta(float theta);
-    void setCharge(float charge);
-    void setD0(float d0);
+    void setPt(double pt);
+    void setEta(double eta);
+    void setPhi(double phi);
+    void setTheta(double theta);
+    void setCharge(double charge);
+    void setD0(double d0);
 
-    float deltaR(float phi1, float eta1, float phi2, float eta2) const;
-    float deltaR(const ParticlePointer otherParticle) const;
-    float deltaR(const PseudoParticlePointer otherParticle) const;
-    float deltaPhi(float phi1, float phi2) const;
-    float deltaEta(float eta1, float eta2) const;
-    bool isWithinDeltaR(float delta_R, const ParticlePointer particle) const;
-    bool isWithinDeltaR(float delta_R, const PseudoParticlePointer particle) const;
+    double deltaR(double phi1, double eta1, double phi2, double eta2) const;
+    double deltaR(const ParticlePointer otherParticle) const;
+    double deltaR(const PseudoParticlePointer otherParticle) const;
+    double deltaPhi(double phi1, double phi2) const;
+    double deltaEta(double eta1, double eta2) const;
+    bool isWithinDeltaR(double delta_R, const ParticlePointer particle) const;
+    bool isWithinDeltaR(double delta_R, const PseudoParticlePointer particle) const;
 };
 
 }
